+++ lib/Nosto/Helper/HtmlMarkupSerializationHelper.php
@@ -62,7 +62,12 @@ class Nosto_Helper_HtmlMarkupSerializationHelper extends Nosto_Helper_AbstractHe
     {
         $spacesStr = str_repeat(' ', $spaces);
         $markup = $spacesStr . self::DIV_START_NOTRANSLATE;
-        $markup .= self::toHtml($object, $key, $spaces + $indent, $indent, self::STYLE_DISPLAY_NONE);
+        $markup .= self::toHtml(
+            $object, Nosto_Helper_SerializationHelper::toSnakeCase($key),
+            $spaces + $indent,
+            $indent,
+            self::STYLE_DISPLAY_NONE
+        );
         $markup .= $spacesStr . self::DIV_END;
         return $markup;
     }
@@ -90,7 +95,7 @@ class Nosto_Helper_HtmlMarkupSerializationHelper extends Nosto_Helper_AbstractHe
         $spacesStr = str_repeat(' ', $spaces);
 
         if ($object instanceof Nosto_Types_MarkupableInterface) {
-            $key = $object->getMarkupKey();
+            $key = Nosto_Helper_SerializationHelper::toSnakeCase($object->getMarkupKey());
         }
 
         //begin block
@@ -98,7 +103,7 @@ class Nosto_Helper_HtmlMarkupSerializationHelper extends Nosto_Helper_AbstractHe
         if ($style) {
             $styleStatement = sprintf(' style="%s"', $style);
         }
-        $classStatement = sprintf(' class="%s"', Nosto_Helper_SerializationHelper::toSnakeCase($key));
+        $classStatement = sprintf(' class="%s"', $key);
         $markup = $spacesStr
             . '<span'
             . $classStatement
@@ -114,10 +119,12 @@ class Nosto_Helper_HtmlMarkupSerializationHelper extends Nosto_Helper_AbstractHe
             $traversable = null;
             if (is_array($object) || $object instanceof Traversable) {
                 $traversable = $object;
+                // Do not convert associative array keys to snake case. It is used for the custom fields
+                $markup .= self::arrayToHtml($object, $key, $spaces, $indent, $traversable, false);
             } elseif (is_object($object)) {
                 $traversable = Nosto_Helper_SerializationHelper::getProperties($object);
+                $markup .= self::arrayToHtml($object, $key, $spaces, $indent, $traversable, true);
             }
-            $markup .= self::arrayToHtml($object, $key, $spaces, $indent, $traversable);
             //end block
             $markup .= $spacesStr . self::SPAN_END . PHP_EOL;
         }
@@ -131,9 +138,10 @@ class Nosto_Helper_HtmlMarkupSerializationHelper extends Nosto_Helper_AbstractHe
      * @param int $spaces
      * @param int $indent
      * @param array|Traversable $traversable the array
+     * @param bool $snakeCaseKey convert the key to snake cases
      * @return string
      */
-    private static function arrayToHtml($object, $key, $spaces, $indent, $traversable)
+    private static function arrayToHtml($object, $key, $spaces, $indent, $traversable, $snakeCaseKey)
     {
         $markup = '';
 
@@ -147,6 +155,10 @@ class Nosto_Helper_HtmlMarkupSerializationHelper extends Nosto_Helper_AbstractHe
                 } else {
                     $childMarkupKey = $key;
                 }
+            }
+
+            if ($snakeCaseKey) {
+                $childMarkupKey = Nosto_Helper_SerializationHelper::toSnakeCase($childMarkupKey);
             }
 
             if ($childValue !== null) {