+++ lib/Nosto/Object/Cart/Cart.php
@@ -49,7 +49,7 @@ class Nosto_Object_Cart_Cart extends Nosto_AbstractObject implements Nosto_Types
     /**
      * @var string URL for restoring cart
      */
-    private $restoreCartUrl;
+    private $restoreLink;
 
     /**
      * @var Nosto_Types_LineItemInterface[] the array of items in the shopping cart
@@ -94,9 +94,9 @@ class Nosto_Object_Cart_Cart extends Nosto_AbstractObject implements Nosto_Types
      *
      * @return string
      */
-    public function getRestoreCartUrl()
+    public function getRestoreLink()
     {
-        return $this->restoreCartUrl;
+        return $this->restoreLink;
     }
 
     /**
@@ -104,9 +104,31 @@ class Nosto_Object_Cart_Cart extends Nosto_AbstractObject implements Nosto_Types
      *
      * @param string $restoreCartUrl
      */
+    public function setRestoreLink($restoreCartUrl)
+    {
+        $this->restoreLink = $restoreCartUrl;
+    }
+
+    /**
+     * Returns the restore cart URL
+     *
+     * @return string
+     * @deprecated use getRestoreLink() instead
+     */
+    public function getRestoreCartUrl()
+    {
+        return $this->getRestoreLink();
+    }
+
+    /**
+     * Sets the restore cart URL
+     *
+     * @param string $restoreCartUrl
+     * @deprecated use setRestoreLink() instead
+     */
     public function setRestoreCartUrl($restoreCartUrl)
     {
-        $this->restoreCartUrl = $restoreCartUrl;
+        $this->setRestoreLink($restoreCartUrl);
     }
 
     /**