+++ lib/Nosto/Object/Cart/LineItem.php
@@ -54,6 +54,11 @@ class Nosto_Object_Cart_LineItem extends Nosto_AbstractObject implements Nosto_T
     private $productId;
 
     /**
+     * @var string the unique identifier of the purchased sku.
+     */
+    private $skuId;
+
+    /**
      * @var int the quantity of the item included in the OrderConfirm.
      */
     private $quantity;
@@ -120,6 +125,24 @@ class Nosto_Object_Cart_LineItem extends Nosto_AbstractObject implements Nosto_T
     public function setProductId($id)
     {
         $this->productId = $id;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getSkuId()
+    {
+        return $this->skuId;
+    }
+
+    /**
+     * Sets the sku ID for the given cart item.
+     *
+     * @param string $id the product ID.
+     */
+    public function setSkuId($id)
+    {
+        $this->skuId = $id;
     }
 
     /**