+++ lib/Nosto/Operation/UninstallAccount.php
@@ -41,28 +41,15 @@
  * An uninstall notification leaves the account as-is on Nosto but revokes all
  * the API tokens.
  */
-class Nosto_Operation_UninstallAccount extends Nosto_Operation_AbstractOperation
+class Nosto_Operation_UninstallAccount extends Nosto_Operation_AbstractAuthenticatedOperation
 {
-    /**
-     * @var Nosto_Types_Signup_AccountInterface Nosto configuration
-     */
-    private $account;
-
-    /**
-     * Constructor.
-     *
-     * @param Nosto_Types_Signup_AccountInterface $account the Nosto configuration object.
-     */
-    public function __construct(Nosto_Types_Signup_AccountInterface $account)
-    {
-        $this->account = $account;
-    }
-
     /**
      * Sends a POST request to delete an account for a store in Nosto
      *
      * @param Nosto_Types_UserInterface $currentUser
      * @return bool if the request was successful.
+     * @throws Nosto_NostoException
+     * @throws Nosto_Request_Http_Exception_AbstractHttpException
      */
     public function delete(Nosto_Types_UserInterface $currentUser)
     {